+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `laravel-queueable-actions` will be documented in this file
 
+## 2.3.0 - 2019-09-24
+
+- Add `make:action` command
+
+## 2.2.0 - 2019-09-04
+
+- Add Laravel 6.0 support
+
+## 2.1.1 - 2019-08-30
+
+- add 'tries' and 'timeout' as queueable properties
+
 ## 2.1.0 - 2019-06-19
 
 - pass the action class name directly to the ActionJob